core: frontend: extract printParam() into parameter.ts
